categories are now working

+++ Services/CategoryExpansesService/CategoryExpansesService.cs
@@ -16,17 +16,17 @@ namespace plusminus.Services.CategoryExpansesService
             _mapper = mapper;
             _context = context;
         }
-        public async Task<ServiceResponse<List<GetCategoryExpansesDto>>> AddCategoryExpanses(AddCategoryExpansesDto newCategoryExpanses)
+        public async Task<ServiceResponse<GetCategoryExpansesDto>> AddCategoryExpanses(AddCategoryExpansesDto newCategoryExpanses, int userId)
         {
-            var serviceResponse = new ServiceResponse<List<GetCategoryExpansesDto>>();
+            var serviceResponse = new ServiceResponse<GetCategoryExpansesDto>();
             try
             {
                 var expansesCategory = _mapper.Map<AddCategoryExpansesDto, CategoryExpenses>(newCategoryExpanses);
+                expansesCategory.UserId = userId;
                 var addedExpanses = await _context.CategoryExpenses.AddAsync(expansesCategory);
                 await _context.SaveChangesAsync();
-
-                var dbCategoryExpanses = await _context.CategoryExpenses.ToListAsync();
-                serviceResponse.Data = dbCategoryExpanses.Select(ci => _mapper.Map<CategoryExpenses, GetCategoryExpansesDto>(ci)).ToList();
+                
+                serviceResponse.Data = _mapper.Map<CategoryExpenses, GetCategoryExpansesDto>(addedExpanses.Entity);
             }
             catch (Exception ex)
             {
@@ -36,19 +36,20 @@ namespace plusminus.Services.CategoryExpansesService
             return serviceResponse;
         }
 
-        public async Task<ServiceResponse<List<GetCategoryExpansesDto>>> DeleteCategoryExpansesById(int id)
+        public async Task<ServiceResponse<int>> DeleteCategoryExpansesById(int id, int userId)
         {
-            var serviceResponse = new ServiceResponse<List<GetCategoryExpansesDto>>();
+            var serviceResponse = new ServiceResponse<int>();
             try
             {
                 var expansesCategory = await _context.CategoryExpenses.FindAsync(id);
-                if (expansesCategory is null) throw new Exception("Категория не найдена");
+                if (expansesCategory is null || expansesCategory.UserId != userId) throw new Exception("Категория не найдена");
+
+                var currentCategoryId = expansesCategory.Id;
 
                 _context.CategoryExpenses.Remove(expansesCategory);
                 await _context.SaveChangesAsync();
-
-                var dbCategoryExpanses = await _context.CategoryExpenses.ToListAsync();
-                serviceResponse.Data = dbCategoryExpanses.Select(ci => _mapper.Map<CategoryExpenses, GetCategoryExpansesDto>(ci)).ToList();
+                
+                serviceResponse.Data = currentCategoryId;
             }
             catch (Exception ex)
             {
@@ -58,13 +59,13 @@ namespace plusminus.Services.CategoryExpansesService
             return serviceResponse;
         }
 
-        public async Task<ServiceResponse<GetCategoryExpansesDto>> UpdateCategoryExpanses(UpdateCategoryExpansesDto updatedCategoryExpanses)
+        public async Task<ServiceResponse<GetCategoryExpansesDto>> UpdateCategoryExpanses(UpdateCategoryExpansesDto updatedCategoryExpanses, int userId)
         {
             var serviceResponse = new ServiceResponse<GetCategoryExpansesDto>();
             try
             {
                 var expansesCategory = await _context.CategoryExpenses.FindAsync(updatedCategoryExpanses.Id);
-                if (expansesCategory is null) throw new Exception("Категория не найдена");
+                if (expansesCategory is null || expansesCategory.UserId != userId) throw new Exception("Категория не найдена");
 
                 expansesCategory.Name = updatedCategoryExpanses.Name;
                 expansesCategory.Color = updatedCategoryExpanses.Color;
